refactor(css): extract CSS rule filtering logic to separate module
Move all CSS rule filtering logic from css.ts to a new css-filter.ts
file in the same directory. This improves code organization by separating
the filtering logic from the CSS configuration itself.

The new css-filter.ts module exports getJavaScriptRulesToDisable() which
is now imported and used by css.ts to maintain the same functionality.
